requirements.txt
backports.csv==1.0.7
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.2.1
django-bootstrap4==0.0.7
django-datetimepicker==3.14
django-import-export==1.1.0
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
facebook-sdk==3.1.0
idna==2.8
jdcal==1.4
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.5.12
psycopg2==2.7.6.1
PyJWT==1.7.1
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
